email-sms/requirements.txt
streamlit
langchain
langchain-openai
langchain-core
pydantic
openai
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib